requirements.txt
# OpinionBalancer KB-Free Dependencies (Ollama Edition)

# orchestrator
langgraph>=0.2
langchain>=0.3
langchain-core>=0.3
pydantic>=2.7
pyyaml>=6.0
rich>=13.7

# LLM client (OpenAI-compatible to Ollama)
langchain-openai>=0.2

# embeddings & metrics
sentence-transformers>=2.7
scikit-learn>=1.4
scipy>=1.12
numpy>=1.26
pandas>=2.2
textstat>=0.7
nltk>=3.9
joblib>=1.4

# additional utilities
requests>=2.31
